asan: reduce resident memory consumption.
When clearing the shadow bytes, use mmap instead memset whenever possible
in order to take benefit from the overcommit of the kernels that support
it. In those kernels, the memory is not allocated before a write occurs,
as a consequence "memset" actually allocates memory, while "mmap" will
only decrease the size of the resident set.

Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>